Level 3 - Terraform Essentials/lab-41_developers-essentials-app-dev-with-secret-manager/requirements.txt

Flask==3.*
google-cloud-secret-manager==2.*

